+++ safety_application/lib/pages/user_page.dart
@@ -1,0 +1,29 @@
+
+
+
+
+// import 'package:flutter/material.dart';
+
+// class UserPage extends StatefulWidget {
+//   const UserPage({super.key});
+
+//   @override
+//   State<UserPage> createState() => _UserPageState();
+// }
+
+// class _UserPageState extends State<UserPage> {
+//   @override
+//   Widget build(BuildContext context) {
+//     return ElevatedButton(
+//       onPressed: _isLoading ? null : _sendPanicAlert,
+//       child: _isLoading ? CircularProgressIndicator() : Text('Panic Button'),
+//       style: ElevatedButton.styleFrom(
+//         // primary: Colors.red,
+//         // onPrimary: Colors.white,
+//         padding: EdgeInsets.symmetric(horizontal: 32, vertical: 16),
+//       ),
+//     );
+//   }
+// }
+    
+// }
